Allow roundtrip dumping of map[interface{}].
- Implement dumping of structs and its subclasses and instances.
- Implement dumping of interface{}.

Remove `types` pointer and use the encoder's loader (for now).

This is an ugly hack. It stores the loader of a type onto the type
itself.

I think the better way is to pass around a context object which will
hold types, and can be used by both Loader and Dumper.

The Dumper can't dump types of nested values because there's no way for
GoType.encode to signal upstream if it sees a nested custom type.